# Fix EMPROFILE=2 when using em++ rather than emcc