--- a/em++.py
+++ b/em++.py
@@ -11,7 +11,7 @@ emcc.run_via_emxx = True
 
 if __name__ == '__main__':
   try:
-    sys.exit(emcc.run(sys.argv))
+    sys.exit(emcc.main(sys.argv))
   except KeyboardInterrupt:
     emcc.logger.warning('KeyboardInterrupt')
     sys.exit(1)